move ethervault/tokenvautl to own folders, addtl refactors from task/utils to regular utils folders, reusabel deploy functions for taikol1 and taikol2
