+++ packages/protocol/tasks/utils.ts
@@ -1,7 +1,5 @@
 import * as fs from "fs";
 import * as log from "./log";
-import { Block, BlockHeader, EthGetProofResponse } from "../test/utils/rpc";
-import RLP from "rlp";
 
 async function deployContract(
     hre: any,
@@ -68,92 +66,8 @@ function getDeployments(_fileName: string) {
     return JSON.parse(`${json}`);
 }
 
-async function getMessageStatusSlot(hre: any, signal: any) {
-    return hre.ethers.utils.solidityKeccak256(
-        ["string", "bytes"],
-        ["MESSAGE_STATUS", signal]
-    );
-}
-
 async function decode(hre: any, type: any, data: any) {
     return hre.ethers.utils.defaultAbiCoder.decode([type], data).toString();
-}
-
-function getSignalSlot(hre: any, sender: any, signal: any) {
-    return hre.ethers.utils.keccak256(
-        hre.ethers.utils.solidityPack(
-            ["string", "address", "bytes32"],
-            ["SIGNAL", sender, signal]
-        )
-    );
-}
-
-const MessageStatus = {
-    NEW: 0,
-    RETRIABLE: 1,
-    DONE: 2,
-    FAILED: 3,
-};
-
-async function getLatestBlockHeader(hre: any) {
-    const block: Block = await hre.ethers.provider.send(
-        "eth_getBlockByNumber",
-        ["latest", false]
-    );
-
-    const logsBloom = block.logsBloom.toString().substring(2);
-
-    const blockHeader: BlockHeader = {
-        parentHash: block.parentHash,
-        ommersHash: block.sha3Uncles,
-        beneficiary: block.miner,
-        stateRoot: block.stateRoot,
-        transactionsRoot: block.transactionsRoot,
-        receiptsRoot: block.receiptsRoot,
-        logsBloom: logsBloom.match(/.{1,64}/g)!.map((s: string) => "0x" + s),
-        difficulty: block.difficulty,
-        height: block.number,
-        gasLimit: block.gasLimit,
-        gasUsed: block.gasUsed,
-        timestamp: block.timestamp,
-        extraData: block.extraData,
-        mixHash: block.mixHash,
-        nonce: block.nonce,
-        baseFeePerGas: block.baseFeePerGas ? parseInt(block.baseFeePerGas) : 0,
-    };
-
-    return { block, blockHeader };
-}
-
-async function getSignalProof(
-    hre: any,
-    contractAddress: string,
-    key: string,
-    blockNumber: number,
-    blockHeader: BlockHeader
-) {
-    const proof: EthGetProofResponse = await hre.ethers.provider.send(
-        "eth_getProof",
-        [contractAddress, [key], blockNumber]
-    );
-
-    // RLP encode the proof together for LibTrieProof to decode
-    const encodedProof = hre.ethers.utils.defaultAbiCoder.encode(
-        ["bytes", "bytes"],
-        [
-            RLP.encode(proof.accountProof),
-            RLP.encode(proof.storageProof[0].proof),
-        ]
-    );
-    // encode the SignalProof struct from LibBridgeSignal
-    const signalProof = hre.ethers.utils.defaultAbiCoder.encode(
-        [
-            "tuple(tuple(bytes32 parentHash, bytes32 ommersHash, address beneficiary, bytes32 stateRoot, bytes32 transactionsRoot, bytes32 receiptsRoot, bytes32[8] logsBloom, uint256 difficulty, uint128 height, uint64 gasLimit, uint64 gasUsed, uint64 timestamp, bytes extraData, bytes32 mixHash, uint64 nonce, uint256 baseFeePerGas) header, bytes proof)",
-        ],
-        [{ header: blockHeader, proof: encodedProof }]
-    );
-
-    return signalProof;
 }
 
 export {
@@ -163,10 +77,5 @@ export {
     getContract,
     saveDeployments,
     getDeployments,
-    getMessageStatusSlot,
-    getSignalSlot,
     decode,
-    MessageStatus,
-    getLatestBlockHeader,
-    getSignalProof,
 };
